backend/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Gong API
requests==2.31.0
httpx==0.26.0

# Vector store and embeddings
pinecone-client==3.0.0
openai==1.12.0
tiktoken>=0.5.2,<0.6.0

# LangChain for RAG
langchain==0.1.6
langchain-openai==0.0.5
langchain-pinecone==0.0.2
langchain-community==0.0.19

# Text processing
nltk==3.8.1

# Configuration and utilities
python-dotenv==1.0.1
pydantic==2.6.0
pydantic-settings==2.1.0

# Async support
aiohttp==3.9.3

# Logging and monitoring
structlog==24.1.0

# Slack integration
slack-sdk==3.27.1

# Testing
pytest>=7.0.0,<8
pytest-asyncio==0.23.4